+++ packages/@expo/metro-config/src/transform-worker/__tests__/collect-dependencies-upstream.test.ts
@@ -902,6 +902,21 @@ it('respects magic comments (webpack) when collecting', () => {
   );
 });
 
+it('respects magic comments (webpack minified) when collecting', () => {
+  const ast = astFromCode(`
+    import(/*webpackIgnore:true*/ "some/async/module").then(foo => {});
+  `);
+  const { dependencies } = collectDependencies(ast, opts);
+  expect(dependencies).toEqual([
+    // Should be empty
+  ]);
+  expect(codeFromAst(ast)).toEqual(
+    comparableCode(`
+      import(/*webpackIgnore:true*/"some/async/module").then(foo => {});
+    `)
+  );
+});
+
 it('respects magic comments when collecting', () => {
   const ast = astFromCode(`
     import(/* @metro-ignore */ "some/async/module").then(foo => {});